reouting_windows refactoring
